users/shopping_cart_service/requirements.txt
Flask==2.2.3
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.5
flask-cors==3.0.10
flask-bcrypt==1.0.1
marshmallow==3.13.0
gunicorn==20.1.0
Flask-Injector==0.14.0